Fix on 32-bit

+++ src/Containers/DenseAxisArray.jl
@@ -26,7 +26,7 @@ end
 
 Base.Array{T,N}(x::DenseAxisArray) where {T,N} = convert(Array{T,N}, x.data)
 
-function Base.hash(d::DenseAxisArray, h::UInt64)
+function Base.hash(d::DenseAxisArray, h::UInt)
     return hash(d.data, hash(d.axes, hash(d.lookup, h)))
 end
